# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  backend_internal_domain_models.BalanceTransaction:
    properties:
      amount:
        type: number
      completed_at:
        type: string
      created_at:
        type: string
      currency:
        type: string
      description:
        type: string
      id:
        type: integer
      payment_details:
        items:
          type: integer
        type: array
      payment_method:
        type: string
      status:
        type: string
      type:
        type: string
      user_id:
        type: integer
    type: object
...
  backend_internal_domain_models.UserBalance:
    properties:
      balance:
        type: number
      currency:
        type: string
      frozen_balance:
        type: number
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  backend_internal_domain_models.UserProfile:
    properties:
      account_status:
        type: string
      bio:
        type: string
      city:
        description: Теперь используем обычные строки, а не указатели
        type: string
      country:
        type: string
      created_at:
        type: string
      email:
        type: string
      google_id:
        type: string
      id:
        type: integer
      is_admin:
        description: Флаг администратора
        type: boolean
      last_seen:
        type: string
      name:
        type: string
      notification_email:
        type: boolean
      phone:
        type: string
      picture_url:
        type: string
      provider:
        type: string
      settings:
        items:
          type: integer
        type: array
      timezone:
        type: string
    type: object
  backend_internal_domain_models.UserProfileUpdate:
    properties:
      bio:
        type: string
      city:
        description: Указатель правильный для обновлений
        type: string
      country:
        type: string
      notification_email:
        type: boolean
      phone:
        type: string
      settings:
        items:
          type: integer
        type: array
      timezone:
        type: string
    type: object
...
  backend_pkg_utils.ErrorResponseSwag:
    properties:
      error:
        example: Описание ошибки
        type: string
    type: object
  backend_pkg_utils.SuccessResponseSwag:
    properties:
      data: {}
      success:
        example: true
        type: boolean
    type: object
...
  internal_proj_payments_handler.StripeWebhookRequest:
    properties:
      created:
        example: 1234567890
        type: integer
      data:
        additionalProperties: true
        type: object
      id:
        example: evt_1234567890
        type: string
      object:
        example: event
        type: string
      type:
        example: checkout.session.completed
        type: string
    type: object
  internal_proj_payments_handler.WebhookResponse:
    properties:
      message:
        example: payments.webhook.processed
        type: string
      status:
        example: success
        type: string
    type: object
...
  internal_proj_users_handler.AdminMessageResponse:
    properties:
      message:
        example: admin.users.success.profile_updated
        type: string
    type: object
  internal_proj_users_handler.AdminStatusUpdateRequest:
    properties:
      status:
        enum:
        - active
        - blocked
        - pending
        example: active
        type: string
    required:
    - status
    type: object
  internal_proj_users_handler.AdminUserListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/backend_internal_domain_models.UserProfile'
        type: array
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      pages:
        example: 10
        type: integer
      total:
        example: 100
        type: integer
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@svetu.rs
    name: API Support
    url: https://svetu.rs/support
  description: API для платформы Sve Tu - маркетплейс объявлений
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://svetu.rs/terms
  title: Sve Tu Marketplace API
  version: "1.0"
paths:
...
  /api/v1/admin/users:
    get:
      consumes:
      - application/json
      description: Returns paginated list of all users in the system
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/internal_proj_users_handler.AdminUserListResponse'
              type: object
        "500":
          description: admin.users.error.fetch_failed
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Get all users (Admin)
      tags:
      - admin-users
  /api/v1/admin/users/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently deletes a user from the system
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User deleted
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/internal_proj_users_handler.AdminMessageResponse'
              type: object
        "400":
          description: admin.users.error.invalid_user_id
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "500":
          description: admin.users.error.delete_failed
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Delete user (Admin)
      tags:
      - admin-users
    get:
      consumes:
      - application/json
      description: Returns detailed user information by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/backend_internal_domain_models.UserProfile'
              type: object
        "400":
          description: admin.users.error.invalid_user_id
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "404":
          description: admin.users.error.user_not_found
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Get user by ID (Admin)
      tags:
      - admin-users
    put:
      consumes:
      - application/json
      description: Updates user profile information by administrator
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/backend_internal_domain_models.UserProfileUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/internal_proj_users_handler.AdminMessageResponse'
              type: object
        "400":
          description: admin.users.error.invalid_user_id or admin.users.error.invalid_format
            or admin.users.error.validation_failed
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "500":
          description: admin.users.error.update_failed
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Update user (Admin)
      tags:
      - admin-users
  /api/v1/admin/users/{id}/balance:
    get:
      consumes:
      - application/json
      description: Returns balance information for a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User balance
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/backend_internal_domain_models.UserBalance'
              type: object
        "400":
          description: admin.balance.error.invalid_user_id
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "500":
          description: admin.balance.error.fetch_balance_failed
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Get user balance (Admin)
      tags:
      - admin-users
  /api/v1/admin/users/{id}/status:
    put:
      consumes:
      - application/json
      description: Updates user status (active, blocked, pending)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status update (active, blocked, pending)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_proj_users_handler.AdminStatusUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Status updated
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/internal_proj_users_handler.AdminMessageResponse'
              type: object
        "400":
          description: admin.users.error.invalid_user_id or admin.users.error.invalid_format
            or admin.users.error.invalid_status
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "500":
          description: admin.users.error.status_update_failed
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Update user status (Admin)
      tags:
      - admin-users
  /api/v1/admin/users/{id}/transactions:
    get:
      consumes:
      - application/json
      description: Returns paginated transaction history for a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - default: 20
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of items to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Transaction history
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/backend_internal_domain_models.BalanceTransaction'
                  type: array
              type: object
        "400":
          description: admin.balance.error.invalid_user_id
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "500":
          description: admin.balance.error.fetch_transactions_failed
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Get user transactions (Admin)
      tags:
      - admin-users
...
  /api/v1/payments/stripe/webhook:
    post:
      consumes:
      - application/json
      description: Processes webhook notifications from Stripe about payment status
      parameters:
      - description: Stripe signature for webhook verification
        in: header
        name: Stripe-Signature
        required: true
        type: string
      - description: Stripe webhook payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_proj_payments_handler.StripeWebhookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Webhook processed successfully
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/internal_proj_payments_handler.WebhookResponse'
              type: object
        "400":
          description: payments.webhook.invalid_signature or payments.webhook.processing_error
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      summary: Process payment webhook
      tags:
      - payments
...
securityDefinitions:
  Bearer:
    description: Bearer token для авторизации
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"